LPS-69400 rename taglib to expando-taglib
