--- modules/apps/foundation/expando/expando-custom-field-ui-taglib/src/main/resources/META-INF/taglib-mappings.properties
@@ -1,1 +1,0 @@
-custom_field_ui=/META-INF/custom-field-ui.tld
