+++ modules/apps/foundation/expando/expando-taglib/src/main/java/com/liferay/expando/taglib/servlet/taglib/CustomAttributeListTag.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.expando.custom.field.ui.taglib.servlet.taglib;
+package com.liferay.expando.taglib.servlet.taglib;
 
-import com.liferay.expando.custom.field.ui.taglib.internal.servlet.ServletContextUtil;
+import com.liferay.expando.taglib.internal.servlet.ServletContextUtil;
 import com.liferay.taglib.util.IncludeTag;
 
 import javax.servlet.http.HttpServletRequest;
@@ -23,7 +23,7 @@ import javax.servlet.jsp.PageContext;
 /**
  * @author Brian Wing Shun Chan
  */
-public class CustomAttributeTag extends IncludeTag {
+public class CustomAttributeListTag extends IncludeTag {
 
 	public void setClassName(String className) {
 		_className = className;
@@ -37,12 +37,12 @@ public class CustomAttributeTag extends IncludeTag {
 		_editable = editable;
 	}
 
-	public void setLabel(boolean label) {
-		_label = label;
+	public void setIgnoreAttributeNames(String ignoreAttributeNames) {
+		_ignoreAttributeNames = ignoreAttributeNames;
 	}
 
-	public void setName(String name) {
-		_name = name;
+	public void setLabel(boolean label) {
+		_label = label;
 	}
 
 	@Override
@@ -57,8 +57,8 @@ public class CustomAttributeTag extends IncludeTag {
 		_className = null;
 		_classPK = 0;
 		_editable = false;
+		_ignoreAttributeNames = null;
 		_label = false;
-		_name = null;
 	}
 
 	@Override
@@ -69,24 +69,27 @@ public class CustomAttributeTag extends IncludeTag {
 	@Override
 	protected void setAttributes(HttpServletRequest request) {
 		request.setAttribute(
-			"custom-field-ui:custom-attribute:className", _className);
+			"liferay-expando:custom-attribute-list:className", _className);
 		request.setAttribute(
-			"custom-field-ui:custom-attribute:classPK",
+			"liferay-expando:custom-attribute-list:classPK",
 			String.valueOf(_classPK));
 		request.setAttribute(
-			"custom-field-ui:custom-attribute:editable",
+			"liferay-expando:custom-attribute-list:editable",
 			String.valueOf(_editable));
 		request.setAttribute(
-			"custom-field-ui:custom-attribute:label", String.valueOf(_label));
-		request.setAttribute("custom-field-ui:custom-attribute:name", _name);
+			"liferay-expando:custom-attribute-list:ignoreAttributeNames",
+			_ignoreAttributeNames);
+		request.setAttribute(
+			"liferay-expando:custom-attribute-list:label",
+			String.valueOf(_label));
 	}
 
-	private static final String _PAGE = "/custom_attribute/page.jsp";
+	private static final String _PAGE = "/custom_attribute_list/page.jsp";
 
 	private String _className;
 	private long _classPK;
 	private boolean _editable;
+	private String _ignoreAttributeNames;
 	private boolean _label;
-	private String _name;
 
 }
