+++ modules/apps/foundation/expando/expando-taglib/src/main/java/com/liferay/expando/taglib/servlet/taglib/CustomAttributeTag.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.expando.custom.field.ui.taglib.servlet.taglib;
+package com.liferay.expando.taglib.servlet.taglib;
 
-import com.liferay.expando.custom.field.ui.taglib.internal.servlet.ServletContextUtil;
+import com.liferay.expando.taglib.internal.servlet.ServletContextUtil;
 import com.liferay.taglib.util.IncludeTag;
 
 import javax.servlet.http.HttpServletRequest;
@@ -23,7 +23,7 @@ import javax.servlet.jsp.PageContext;
 /**
  * @author Brian Wing Shun Chan
  */
-public class CustomAttributeListTag extends IncludeTag {
+public class CustomAttributeTag extends IncludeTag {
 
 	public void setClassName(String className) {
 		_className = className;
@@ -37,12 +37,12 @@ public class CustomAttributeListTag extends IncludeTag {
 		_editable = editable;
 	}
 
-	public void setIgnoreAttributeNames(String ignoreAttributeNames) {
-		_ignoreAttributeNames = ignoreAttributeNames;
-	}
-
 	public void setLabel(boolean label) {
 		_label = label;
+	}
+
+	public void setName(String name) {
+		_name = name;
 	}
 
 	@Override
@@ -57,8 +57,8 @@ public class CustomAttributeListTag extends IncludeTag {
 		_className = null;
 		_classPK = 0;
 		_editable = false;
-		_ignoreAttributeNames = null;
 		_label = false;
+		_name = null;
 	}
 
 	@Override
@@ -69,27 +69,24 @@ public class CustomAttributeListTag extends IncludeTag {
 	@Override
 	protected void setAttributes(HttpServletRequest request) {
 		request.setAttribute(
-			"custom-field-ui:custom-attribute-list:className", _className);
+			"liferay-expando:custom-attribute:className", _className);
 		request.setAttribute(
-			"custom-field-ui:custom-attribute-list:classPK",
+			"liferay-expando:custom-attribute:classPK",
 			String.valueOf(_classPK));
 		request.setAttribute(
-			"custom-field-ui:custom-attribute-list:editable",
+			"liferay-expando:custom-attribute:editable",
 			String.valueOf(_editable));
 		request.setAttribute(
-			"custom-field-ui:custom-attribute-list:ignoreAttributeNames",
-			_ignoreAttributeNames);
-		request.setAttribute(
-			"custom-field-ui:custom-attribute-list:label",
-			String.valueOf(_label));
+			"liferay-expando:custom-attribute:label", String.valueOf(_label));
+		request.setAttribute("liferay-expando:custom-attribute:name", _name);
 	}
 
-	private static final String _PAGE = "/custom_attribute_list/page.jsp";
+	private static final String _PAGE = "/custom_attribute/page.jsp";
 
 	private String _className;
 	private long _classPK;
 	private boolean _editable;
-	private String _ignoreAttributeNames;
 	private boolean _label;
+	private String _name;
 
 }
