+++ modules/apps/foundation/expando/expando-taglib/src/main/resources/META-INF/liferay-expando.tld
@@ -8,11 +8,11 @@
 >
 	<description>This tag is used for display expando custom fields.</description>
 	<tlib-version>1.0</tlib-version>
-	<short-name>custom-field-ui</short-name>
-	<uri>http://liferay.com/tld/custom_field_ui</uri>
+	<short-name>liferay-expando</short-name>
+	<uri>http://liferay.com/tld/expando</uri>
 	<tag>
 		<name>custom-attribute</name>
-		<tag-class>com.liferay.expando.custom.field.ui.taglib.servlet.taglib.CustomAttributeTag</tag-class>
+		<tag-class>com.liferay.expando.taglib.servlet.taglib.CustomAttributeTag</tag-class>
 		<body-content>JSP</body-content>
 		<attribute>
 			<name>className</name>
@@ -42,7 +42,7 @@
 	</tag>
 	<tag>
 		<name>custom-attribute-list</name>
-		<tag-class>com.liferay.expando.custom.field.ui.taglib.servlet.taglib.CustomAttributeListTag</tag-class>
+		<tag-class>com.liferay.expando.taglib.servlet.taglib.CustomAttributeListTag</tag-class>
 		<body-content>JSP</body-content>
 		<attribute>
 			<name>className</name>
@@ -72,7 +72,7 @@
 	</tag>
 	<tag>
 		<name>custom-attributes-available</name>
-		<tag-class>com.liferay.expando.custom.field.ui.taglib.servlet.taglib.CustomAttributesAvailableTag</tag-class>
+		<tag-class>com.liferay.expando.taglib.servlet.taglib.CustomAttributesAvailableTag</tag-class>
 		<body-content>JSP</body-content>
 		<attribute>
 			<name>className</name>
