+++ modules/apps/foundation/expando/expando-taglib/src/main/resources/META-INF/resources/custom_attribute/init.jsp
