+++ modules/apps/foundation/expando/expando-taglib/src/main/resources/META-INF/taglib-mappings.properties
@@ -1,0 +1,1 @@
+liferay_expando=/META-INF/liferay-expando.tld
